all analysers should inherit from analyser
